Split identify in two modules
